augmentations for scaling images on single axis
Adds augmentations for scaling in just the X or Y dimensions. This is useful for working with video, which may have images that are off in terms of aspect ratio (particularly animation and video games, where the aspect ratio can be fudged significantly without looking too weird), though as I note in the code the individual image pairs should be consistent in this regard between the low quality and GT images. The scaling ratio ranges are intentionally large to cover a range of ratio disparities.